Use integer vector for getWindowSize
Change getWindowSize() return type from glm::vec2 to glm::ivec2 in GFXEngine/Game.h to match the m_windowSize member. This ensures the window size is returned as integer pixel coordinates and avoids implicit float/int mismatches.

+++ GFXEngine/Game.h
@@ -19,7 +19,7 @@ namespace GFXEngine {
 			virtual void onRender(Graphics::Renderer& renderer, uint32_t imageIndex) = 0;
 			virtual void onDestroy(Graphics::Renderer& renderer) = 0;
 
-			glm::vec2 getWindowSize() const { return m_windowSize; }
+			glm::ivec2 getWindowSize() const { return m_windowSize; }
 
 		private:
 			GLFWwindow* m_window = nullptr;
